fix: Revisit entrypoint
