--- .vscode/settings.json
@@ -1,14 +1,0 @@
-{
-  "ruff.enable": true,
-  "ruff.configuration": "pyproject.toml",
-  "python.testing.pytestEnabled": true,
-  "python.testing.pytestArgs": [
-    "tests"
-  ],
-  "githubIssues.issueBranchTitle": "issues/${issueNumber}-${issueTitle}",
-  "pylint.importStrategy": "fromEnvironment",
-  "pylint.args": [
-    "--rcfile=pyproject.toml"
-  ],
-
-}
