+++ anta/cli/_main.py
@@ -69,7 +69,3 @@ def cli() -> None:
             logger,
         )
         sys.exit(ExitCode.INTERNAL_ERROR)
-
-
-if __name__ == "__main__":
-    anta()
