+++ tests/units/cli/debug/test_commands.py
@@ -9,7 +9,7 @@ from typing import TYPE_CHECKING, Literal
 
 import pytest
 
-from anta.cli import anta
+from anta.cli._main import anta
 from anta.cli.utils import ExitCode
 
 if TYPE_CHECKING:
